backend/requirements.txt
Django==6.0.2
psycopg2-binary==2.9.9
python-decouple==3.8
djangorestframework==3.14.0
django-cors-headers==4.3.0
